Create numero position
