+++ lib/geojson.js
@@ -3,7 +3,7 @@ import computeConcave from '@turf/concave'
 import {point} from '@turf/helpers'
 import {randomColor} from 'randomcolor'
 
-import {getStatus, getName, hasPosition, getNumeroPosition, getPosition} from './bal/item'
+import {getStatus, getName, hasPosition, getNumeroPosition, getPosition, getNumeroPositions} from './bal/item'
 
 function getCoordinates(position) { // Converge /explore and BAL models
   const {coords, coordinates} = position
@@ -76,7 +76,7 @@ export function featureToPosition(feature) {
   return {
     _id: feature.id,
     coords: feature.geometry.coordinates,
-    type: 'entrée',
+    type: feature.type || 'entrée',
     source: [],
     dateMAJ: null
   }
@@ -182,9 +182,10 @@ export function numerosToGeoJson(numeros) {
 }
 
 export function numeroPositionsToGeoJson(numero) {
+  const positions = getNumeroPositions(numero)
   return {
     type: 'FeatureCollection',
-    features: numero.positions.map(position => positionToFeature(position, numero))
+    features: positions.map(position => positionToFeature(position, numero))
   }
 }
